Fix intermittently failing test
There is a possibility that the client receives data immediately as the
HTTP handshake completes. In this scenario, the new websocket data can
be in the same chunk as the header. Previously this was sometimes
getting lost (causing tests to be tempramental). This fixes the issue,
though it is, sadly, pretty hacky.